luajs: replace luaJS_eval_js with a variant using the new JSC API
The original JavaScriptCore API has been deprecated in favour of a new
GObject-based one. Add a luajs_eval_js function using the new API along
with a necessary helper function, replace all the calls to the old
variant and remove that one.

No functional change intended, except for the new luajs_pushvalue
function handling objects with empty strings or larger-than-int integers
as keys correctly and a different wording of an error message.