page: allow for returning functions from eval_js
Absorb the functionality handled by wrap_js() until now into eval_js().
This is motivated by the fact that in the new JSC API there doesn't seem
to exist a function to wrap a JavaScript snippet as a function and
instead a JS code which generates the function in full must be evaluated
anyway.

Thus, if the last value generated by the evaluated code is a JavaScript
function, it is returned to Lua and can be called directly. wrap_js()
has been updated to serve as a mere wrapper for backwards compatibility
(despite this function never being publicly documented). All it does is
it wraps the provided JS snippet string to evaluate to an anonymous
function and passes this to eval_js().

Also document the return value for page:eval_js().