page: use the new JSC API for wrap_js
Update the code relevant to the page:wrap_js() function to use the
GObject-style JavaScriptCore API instead of the deprecated old one.

This brings some changes to how this function is used: before, the
function took the JavaScript function body as the first parameter and a
sequence of its parameter names as the second, whereas now, the entire
function definition must be part of the JS snippet which must
additionally end by a noop statement of the function object, no other
parameters are passed to the Lua function. Hopefully this should be fine
since the function is not documented in the public API and is currently
in use only in lib/select_wm.lua, where its use has been appropriately
updated -- see this file for an illustration of the above description.

Also, the function now checks for JavaScript evaluation errors and
properly propagates them to Lua.