select_wm: do not unnecessarily re-evaluate client_rects
One of the advantages documented for JSObjectMakeFunction of the
deprecated JSC API is that it's useful for repeated invocations of the
same code as it avoids reparsing it every time [1]. This likely also
holds for the new API which is now used for page:wrap_js(). While the
speed improvement is unlikely to be noticeable in most cases (if in
any), it still seems to be the correct/more logical approach code-wise
to evaluate it only once before the for loop rather than for every
selectable element of which there can be quite a few.

[1] https://developer.apple.com/documentation/javascriptcore/jsobjectmakefunction(_:_:_:_:_:_:_:_:)