webview: ignore zoom_text_only setting
All settings are loaded in no particular order. Whenever the
zoom_text_only setting was applied after the zoom_level setting,
the zoom_level setting was set to 0.5. This happened occasionally.

This commit deactivates the zoom_text_only property until there's
a way to load the settings tables in a controlled sequence.

Fixes: #1086 #1013 #922