config: remove hide_scrollbars from the default rc.lua
Hiding the scrollbars causes rendering issues with smooth scrolling --
in particular, the scrolling doesn't often take effect until some other
rendering is forced, such as changing a link color by hovering over it,
invoking the follow hints or selecting some text which causes it to get
highlighted.

Other approaches to hide the bars, such as using "display: none"
instead, lead to the same problem. It is possible that it will be
possible to restore this eventually using the standardized scrollbar
properties (scrollbar-width) provided they won't have the same
shortcoming as this WebKit-specific one.

Fixes: https://github.com/luakit/luakit/issues/1007
Link: https://github.com/luakit/luakit/issues/1081#issuecomment-2081564587