webview: default HW acceleration policy to always
The on-demand default has been reported to cause rendering issues (such
as the view not being redrawn on scroll while hovering over it reveals
that the scroll did happen by recognizing the new positions of clickable
elements for instance) on at least some systems, although it doesn't
seem to be always reproducible. The same problem has been observed with
the WebKit MiniBrowser too (although also unreliably), so it is likely a
WebKit issue.

WebKit itself seems to default to always too plus the on-demand option
is not available in webkitgtk-6.0 so this change should be sensible in
any case.

Fixes: https://github.com/luakit/luakit/issues/1081